% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specc.R
\name{adjacency}
\alias{adjacency}
\title{Computes adjacency matrix given similarity matrix. Each observation is connected
to its K-nearest neighbors and the similarity measures as provided by the
similarity matrix are used as edge weights.}
\usage{
adjacency(distance.mat, similarity.mat, K)
}
\arguments{
\item{distance.mat}{Matrix of pairwise distances between observations}

\item{similarity.mat}{Matrix of pairwise similarities between observations}

\item{K}{Number of nearest neighbors to connect each observation to}
}
\value{
Adjacency matrix
}
\description{
Computes adjacency matrix given similarity matrix. Each observation is connected
to its K-nearest neighbors and the similarity measures as provided by the
similarity matrix are used as edge weights.
}
\examples{
distMat <- pairwiseDistance(iris[, 1:4])
similarityMat <- similarity(iris[, 1:4])
adjacency(distMat, similarityMat, 3)
}
