% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specc.R
\name{degree}
\alias{degree}
\title{Compute matrix with degree of each vertex on the diagonal}
\usage{
degree(adjacency.mat)
}
\arguments{
\item{adjacency.mat}{Adjacency matrix}
}
\value{
Matrix with degrees of vertex i on [ i, i ]
}
\description{
Compute matrix with degree of each vertex on the diagonal
}
\examples{
adjacency.mat = adjacency(pairwiseDistance(iris[1:50,1:4]),
similarity(iris[,1:4]), 2)
degree(adjacency.mat)
}
