% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{HCLUST}
\alias{HCLUST}
\title{Perform Hierarchical Clustering}
\usage{
HCLUST(data.matrix, K)
}
\arguments{
\item{data.matrix}{input data}

\item{K}{number of clusters}
}
\value{
a vector of cluster assignments
}
\description{
Perform agglomerative hierarchical clustering on a given data matrix and
create K clusters
}
\examples{
HCLUST(iris[,1:4], 3)
}
