% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specc.R
\name{laplacian}
\alias{laplacian}
\title{Calculates graph Laplacian given adjacency and degree matrices}
\usage{
laplacian(adjacency.mat, degree.mat)
}
\arguments{
\item{adjacency.mat}{Adjacency matrix of graph}

\item{degree.mat}{Matrix with degrees of vertices on diagonal}
}
\value{
Laplacian matrix = degree - adjacency
}
\description{
Calculates graph Laplacian given adjacency and degree matrices
}
\examples{
adjacency.mat <- adjacency(pairwiseDistance(iris[1:50,1:4]), similarity(iris[,1:4],
2)
degree.mat <- degree(adjacency.mat)
laplacian.mat <- laplacian(adjacency.mat, degree.mat)
}
