% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specc.R
\name{nearestNeighbors}
\alias{nearestNeighbors}
\title{Computes K-nearest neighbors of given observation}
\usage{
nearestNeighbors(distance.mat, point, K)
}
\arguments{
\item{distance.mat}{Matrix of pairwise distances of observations}

\item{point}{Given observation to find nearest neighbors of}

\item{K}{number of nearest neighbors}
}
\value{
Vector of nearest neighbors
}
\description{
Computes K-nearest neighbors of given observation
}
\examples{
distMat <- pairwiseDistance(iris.dt[, 1:4])
neighborsVec <- nearestNeighbors(distMat, 3, 2)
}
