% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specc.R
\name{similarity}
\alias{similarity}
\title{Computes radial-kernel gram matrix to use as similarity matrix with scale
parameter = 1 and Euclidean distance}
\usage{
similarity(data.matrix)
}
\arguments{
\item{data.matrix}{Matrix of observations}
}
\value{
Similarity matrix between observations
}
\description{
Computes radial-kernel gram matrix to use as similarity matrix with scale
parameter = 1 and Euclidean distance
}
\examples{
similarity(iris[1:50,1:4])
}
