% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specc.R
\name{SPECC}
\alias{SPECC}
\title{Performs Spectral Clustering}
\usage{
SPECC(data.dt, numClust, numEigen, numNeighbors)
}
\arguments{
\item{data.dt}{Data table of observations}

\item{numClust}{Desired number of clusters}

\item{numEigen}{Desired number of eigenvalues used}

\item{numNeighbors}{Number of neighbors to choose in K-nearest neighbor
computation}
}
\value{
Vector of cluster assignments.
}
\description{
Performs spectral clustering using K-nearest neighbors (where K is passed in
as a parameter) using Euclidean distances. Uses K-means clustering on the
eigenvectors.
}
\examples{
library(data.table)
set.seed(1)
halfcircle <- function(r, center = c(0, 0), class, sign, N=150, noise=0.5) {
 angle <- runif(N, 0, pi)
 rad <- rnorm(N, r, noise)
 data.table(
   V1 = rad * cos(angle) + center[1],
   V2 = sign * rad * sin(angle) + center[2]
   )
}
X.dt <- rbind(
 halfcircle(4, c(0, 0), 1, 1),
 halfcircle(4, c(4, 2), 2, -1))

result <- SPECC(X.dt, 2, 5, 5)
}
