% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R, R/specc.R
\name{pairwiseDistance}
\alias{pairwiseDistance}
\title{Compute matrix of pairwise Euclidean distance between observations}
\usage{
pairwiseDistance(data.matrix)

pairwiseDistance(data.matrix)
}
\arguments{
\item{data.matrix}{Data matrix of n observations}
}
\value{
nxn matrix with Euclidean distance between observation a and
observation b located at [ a,b ] and [ b,a ]

nxn matrix with Euclidean distance between observation a and
observation b located at [ a,b ] and [ b,a ]
}
\description{
Compute matrix of pairwise Euclidean distance between observations

Compute matrix of pairwise Euclidean distance between observations
}
\examples{
pairwiseDistance(as.matrix(iris[,1:4]))
pairwiseDistance(as.matrix(iris[1:50,1:4]))
}
